feat: Update HTML lang attribute to include data-theme in index.html
The code changes modify the HTML lang attribute in the index.html file to include the data-theme attribute. This change allows for the implementation of a theme switcher functionality in the blog app. By adding the data-theme attribute, the app can dynamically change the theme based on user preferences.

ADD daisy ui on the interface