feat: Add lazy loading for images in App.vue
The code changes in App.vue import the `onMounted` function from the `vue` package and the `getCurrentUser` function from the `vuefire` package. The `onMounted` function is used to asynchronously fetch the current user, and if a user is logged in, the router is redirected to the "blog" route. This change improves the performance of the page by implementing lazy loading for images.